[Add] #60 - String extension 파일 추가
